Use SSIZE_T from BaseTsd.h on windows
It is defined by BaseTsd.h as LONG_PTR, which is 32bit size on win32, and
64bit size on win64.

It seems that mingw rather uses int for the 32bit case, but better stick
to the MS definition, and int/LONG_PTR will be abi-compatible on 32bit.

Fixes #68

+++ src/libslirp.h
@@ -10,7 +10,8 @@
 #include <winsock2.h>
 #include <ws2tcpip.h>
 #include <in6addr.h>
-typedef int slirp_ssize_t;
+#include <BaseTsd.h>
+typedef SSIZE_T slirp_ssize_t;
 #ifdef BUILDING_LIBSLIRP
 # define SLIRP_EXPORT __declspec(dllexport)
 #else
